environment fix: all mentors other than clint were not functional, now fixed
Updated docker image for classifier to fix issues where inference
did not work with any mentor other than clint.

In the new version, pip is used to install all packages
instead of conda, and the package versions are locked specifically.

Another important difference is that the theano package is no longer
installed. Somehow installing theano, triggers a change where the
inference engine uses theano instead of tensorflow, and that
may be related to the problems we were seeing.