client can load arbitrary mentors
Client has had a hard-coded list of mentors with their initial videos and data,
and has refused to load data for any mentor not in that list.

This is a halfway fix for that problem: retains old hard-coded data
but allows unknown mentors.

A limitation for now is that client will not have name and title for an unknown mentor